Restructure the PUT functions to allow some simpler entry points
External programs may wish to use these two new functions in order
to create or update events within the DAViCal database.  While this
probably is not a completely stable API, the simple_write_resource()
function in particular will be retained significantly after a real
API is available, and possibly forever.