[sync-remote] More attempts to handle Google.
It seems that on PUT Google might truncate the filename, so it will
send us back a Location header we need to follow before we can
getr the ETag.  And there might be one more hoop to jump through
to complete this process.

This would be better handled if we locally use the VEVENT UID as
the invariant property between local and remote, rather than the
'filename' part of the path, but we'd still face the issue of
having to follow the Location: header to get the ETag for the
event we just PUT to the remote server.