Move from extract.pl to standard xgettext
So far we have used extract.pl which originated in Horde to generate the
PO files. This process took a long time to run.

But xgettext is able to handle PHP by itself.

In the source translate() and i18n() are used instead of _() therefore
we have to pass a keyword parameter.

From now on Translators: is the keyword to provide content to the
translators on Transifex.

Conflicts:

	scripts/po/extract.pl